add/edit-merch-CRUD-functionality
