+++ client/components/ListMerch.tsx
@@ -26,19 +26,6 @@ const ListMerch = () => {
     }
   }
 
-  // //get merch function
-  // const getMerch = async (id: number) => {
-  //   try {
-  //     const response = await fetch(`http://localhost:3001/admin/admin-merch/${id}`);
-  //     const jsonData = await response.json();
-
-  //     setMerches(jsonData);
-  //   } catch (err) {
-  //     console.error(getErrorMessage(err));
-  //   }
-  // }
-
-  //get all merch function
   const getAllMerches = async () => {
     try {
       const response = await fetch("http://localhost:3001/admin/admin-merch");
@@ -65,13 +52,6 @@ const ListMerch = () => {
         </tr>
       </thead>
       <tbody>
-        {/*
-        <tr>
-          <td>John</td>
-          <td>Doe</td>
-          <td>john@example.com</td>
-        </tr> 
-        */}
         {merches.map(merch => (
           <tr key={merch.id}>
             <td className="border px-4 py-2">{merch.name}</td>
@@ -79,12 +59,6 @@ const ListMerch = () => {
             {/* <td className="border px-4 py-2">{merch.image}</td> */}
             <td className="border px-4 py-2">P {merch.price}</td>
             <td>
-              {/* <a 
-                className="text-white bg-blue-700 hover:bg-blue-800 focus:ring-4 focus:ring-blue-300 font-medium rounded-lg text-sm px-5 py-2.5 dark:bg-blue-600 dark:hover:bg-blue-700 focus:outline-none dark:focus:ring-blue-800"
-                href="http://localhost:3000/admin/admin-editMerch"
-              >
-                Edit
-              </a> */}
               <EditMerch merch={merch} />
             </td>
             <td>
