--- client/pages/admin/admin-editMerch.tsx
@@ -1,13 +1,0 @@
-import React, { Fragment, useEffect } from 'react';
-
-//components
-
-import EditMerch from '@/components/EditMerch';
-
-export default function AdminMerch() {
-  return (
-    <div className="container">
-      <EditMerch />
-    </div>
-  );
-}
